Fix to unify to one AbstractAuthCommand.
